Crash Fix, Rarity, and Mod Config
Darklights Stone was crashing if you upgraded again at max level. Also fixed an issue with how it was not rendering the text in places.

Adjusted Adventurer's Text to be more clear

Slightly reworked Cat's Pajamas:
Now only grants 9 HP on entering a rest site but grants gold equal to any healing received (Like resting at a rest site). Added a check to avoid an endless loop with Bloody Idol

Attempted to add LegacyMode for Nik the GOAT and a new version that was more balanced. It appears to be stuck in legacy mode. Still working on it.