Sound Updates, minor tweaks
Updated some of the sound effects on the relics.

You can now right click the relics any time in a run instead of only in combat! :)

Changed Ninja Puppy's Mask to apply 3 instances of 1 constrict instead of 1 instead of 2 constrict. This should *destroy* artifact.

Changed HangInThere to give 4 hp instead of 3 hp

Greener Grass now pulses when it's waiting to be used and stops when it's used.

Removed double registered RatRace.