Bug Fixes
Bubble Burster was not triggering on Event bosses.

Burning Bridge now is limited to non-use on the card so it won't go infinitely. This change also includes a new compounding Hp cost each use of the relic to each time you use instead.

Tough Cookie was not triggering properly in all elite rooms. that should be fixed now.