Minor Update
Added in the missing bonus from Bee's Knees but also buffed it in the chance you'll trigger it, making it worth it.

Changed the text of Knuckle Down to be more inline with vanilla:
      "If you play 3 Skills in a turn, remove a random debuff stack from yourself."
      "If you play 3 Skills in a turn, remove a random debuff from yourself."

It still works the same, just the text was updated

Added One For the Road
Common Relic
When you draw a power card, draw a card.